requirements.txt
numpy
pandas
python-dotenv
pytz
requests
urllib3
alpaca-trade-api
polygon-api-client
tqdm